# docs/api.yaml
openapi: 3.0.4
info:
  title: Detective API
  version: v1
paths:
...
  /api/v1/persons:
    get:
      tags:
        - Person
      parameters:
        - name: Sex
          in: query
          schema:
            $ref: '#/components/schemas/SexDto'
        - name: FullName
          in: query
          schema:
            maxLength: 100
            type: string
        - name: MinBirthDate
          in: query
          schema:
            type: string
            format: date
        - name: MaxBirthDate
          in: query
          schema:
            type: string
            format: date
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonDto'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '503':
          description: Внутренний сервис не доступен
    post:
      tags:
        - Person
      parameters:
        - name: Sex
          in: query
          schema:
            $ref: '#/components/schemas/SexDto'
        - name: FullName
          in: query
          schema:
            maxLength: 100
            type: string
        - name: BirthDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: uuid
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '503':
          description: Внутренний сервис не доступен
  '/api/v1/persons/{personId}':
    get:
      tags:
        - Person
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonFullDto'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
    delete:
      tags:
        - Person
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
    patch:
      tags:
        - Person
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sex
          in: query
          schema:
            $ref: '#/components/schemas/SexDto'
        - name: fullName
          in: query
          schema:
            type: string
        - name: birthDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '204':
          description: No Content
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
...
  /api/v1/persons/relationships:
    post:
      tags:
        - Person
      parameters:
        - name: person1Id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: person2Id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/RelationshipTypeDto'
      responses:
        '204':
          description: No Content
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
    delete:
      tags:
        - Person
      parameters:
        - name: person1Id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: person2Id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '503':
          description: Внутренний сервис не доступен
...
  '/api/v1/persons/contacts/{contactId}':
    delete:
      tags:
        - Person
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
...
  '/api/v1/persons/documents/{documentId}':
    delete:
      tags:
        - Person
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
...
  '/api/v1/persons/property/{propertyId}':
    delete:
      tags:
        - Person
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
...
  '/api/v1/persons/characteristic/{characteristicId}':
    delete:
      tags:
        - Person
      parameters:
        - name: characteristicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
    get:
      tags:
        - Person
      parameters:
        - name: characteristicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacteristicDto'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Внутренний сервис не доступен
...
components:
  schemas:
...
    CharacteristicDto:
      required:
        - appearance
        - authorUsername
        - medicalConditions
        - personality
      type: object
      properties:
        id:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        authorUsername:
          type: string
          nullable: true
        appearance:
          type: string
          nullable: true
        personality:
          type: string
          nullable: true
        medicalConditions:
          type: string
          nullable: true
      additionalProperties: false
    ContactDto:
      required:
        - info
      type: object
      properties:
        id:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ContactTypeDto'
        info:
          type: string
          nullable: true
      additionalProperties: false
    ContactTypeDto:
      enum:
        - Phone
        - Email
        - Telegram
        - Vk
      type: string
...
    DocumentDto:
      required:
        - payload
      type: object
      properties:
        id:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/DocumentTypeDto'
        payload:
          type: string
          nullable: true
      additionalProperties: false
    DocumentTypeDto:
      enum:
        - Passport
        - FakePassport
        - DriverLicense
      type: string
...
    PersonDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sex:
          $ref: '#/components/schemas/SexDto'
        fullName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
          nullable: true
      additionalProperties: false
    PersonFullDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sex:
          $ref: '#/components/schemas/SexDto'
        fullName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
          nullable: true
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ContactDto'
          nullable: true
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentDto'
          nullable: true
        properties:
          type: array
          items:
            $ref: '#/components/schemas/PropertyDto'
          nullable: true
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/CharacteristicDto'
          nullable: true
      additionalProperties: false
    PropertyDto:
      required:
        - type
      type: object
      properties:
        id:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        type:
          type: string
          nullable: true
        cost:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
...
    RelationshipTypeDto:
      enum:
        - Parent
        - StepParent
        - Child
        - StepChild
        - Friend
        - Spouse
        - ExSpouse
        - Sibling
        - StepSiblings
        - Employer
        - Employee
        - Colleague
        - Mentor
        - Protege
        - NoMore
      type: string
    SexDto:
      enum:
        - Male
        - Female
      type: string
...
  securitySchemes:
    Bearer:
      type: http
      description: Please enter a valid token
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]